Initial commit
Most functionality from last spring's `scatdb` project preserved,
save for the Telegram bot integration. TODO:
- fix lemmatization of 4th class verbs (previously Telegram-dependent);
- add lemmatizer classes for the remaining parts of speech.

The main module is now named `converter` and fired from the command
line. It may run in 3 modes, w/ output in
- TSV (not further action necessary, save for output format),
- XML (TODO: add post-processor to collapse names and punctuation),
- PKL (TODO: not yet implemented).

The PKL (a simple shelve DB will suffice) output is supposed to be
the basis of the second main module, `annotator`, already implemented
in autumn of 2018 but requiring additional support for homonymous MSDs.