Overhaul number processing
A separate module was added for handling numeric tokens, including
the translation to decimal and normalization. `reg` values as well
as lemmata will now hold decimal digits.

They are also included in the `<num>` node value tag count in the
postprocessing script.

The distinction between cardinals and ordinals was preserved, i.e.
- Д# -> 4, no morphology.
- $ЗКЕ#ГО -> 7025-ыи, all morphology intact.